Move pants_env_vars.py to be under bin/
[ci skip-rust]

[ci skip-build-wheels]
